📝 chore(.gitignore): add IntelliJ IDEA, Eclipse, NetBeans, VS Code, and Mac OS files to ignore Added .gitignore files for IntelliJ IDEA, Eclipse, NetBeans, and VS Code to ignore their default files and directories. Also added Mac OS files to ignore. This improves the cleanliness of the repository and prevents unnecessary files from being committed.
📝 chore(.idea): add IntelliJ IDEA project files to ignore
Added .gitignore files for IntelliJ IDEA project files to ignore their default files and directories. This improves the cleanliness of the repository and prevents unnecessary files from being committed.

📝 chore(.idea): add IntelliJ IDEA dictionaries and inspection profiles
Added IntelliJ IDEA dictionaries and inspection profiles to the repository. This improves the consistency of the project and allows for better code quality.

📝 chore(java-swing.iml): add Java module file
Added Java module file to the repository.